ghana-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
torch==2.3.0
scikit-learn==1.4.2
pandas==2.2.2
numpy==1.26.4
groq==0.9.0
google-generativeai==0.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1
reportlab==4.2.2
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
apscheduler==3.10.4
aiofiles==23.2.1
python-multipart==0.0.9
